requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.4
django-auto-logout==0.5.0
django-axes==5.33.0
django-environ==0.9.0
django-graphos==0.3.41
django-ipware==4.0.2
django-qr-code==3.0.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
fpdf==1.7.2
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
PyPDF2==1.27.12
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
segno==1.5.2
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
